Simplify rule generation for deriving extensions
Previously, the attributes, rules, and the generators would all be
created separately and then composed. That's a bit error prone as there
too many places where one needs to update the codebase to add a new type
of deriving plugin. Now one simply needs to specify the type, and the
various rule generators, and code expanders. It's easier to get this all
right in one place.

Signed-off-by: Rudi Grinberg <rudi.grinberg@gmail.com>